Full stack (final project)
Information present online comes in a wide range of formats and encompasses many different technologies. Very often these technologies intersect, and data present in one technology is used in another. Project goals The project encompasses this nature of modern data storage. In this project you are required to achieve the following goals 1. Make a relational database that uses at least 3 tables. 2. Has scripting present which can execute database queries. 3. Has scripting which obtains information from a public API.  4. Provides a frontend for users that includes at least a single web page. 

This is a recommendation marketing web application that works according to the state of the climate, and it was developed using Php, JSON, HTML, and CSS, and connected to a database that was created using PhpMyAdmin.