AI Ethics-Ethical Audit & Evaluation
This was the (Part 2) of my graduation's final project in the MSc In Information Systems, (Part 1 was the ML project), I got an (A-) for this report.
Here are the requirements:

Overview of ethical issues in AI. Write a brief introduction to ethical issues in artificial intelligence. Include an overview of ethical issues in AI generally. Focus then on one domain/application area or technology that you are interested in. Describe ethical issues or risks associated with this.
Review current approaches to ensuring ethical AI. Briefly summarise and critique different kinds of solutions. Identify those that are relevant to your domain area. Review 1-3 key works/projects/approaches that sought to address ethical issues. If there are no solutions to ethical issues in your chosen domain, review solutions in a different domain that you think could be used to address issues in your selected area. 
Review approaches to ethical evaluations of systems with a focus on ethical audits/evaluations. Identify any that are relevant to your domain. Address whether there is a need for standard ethical audit of systems in your particular domain. If so, explain why and discuss the specific requirements of that domain that need to be addressed. 
Select an approach to evaluating an example system in your domain as an external auditor/evaluator. You may select the ALTAI audit and tailor it to your domain. Design an ethical audit that could be used by a company/organisation to evaluate their systems. Be as detailed and domain specific as possible. The point here is to go beyond high-level principles. 
Compile a report including the sections above and detailing your evaluation/audit design including. The paper should be 6-8 pages with additional information such as audit documents detailed in an appendix. 
Create accompanying video presentation (slides and audio) for 3 minutes max, on your project. 

I came up with the subject of (Body Worn Cameras with an embedded FRT, Ethical Audit and evaluation). Please read the full report in the attached file.